# Fix `Quests only` unresponsiveness with large Pokéstop payload on eve…